requirements.txt
anyio==4.6.2.post1
asgiref==3.3.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
certifi==2024.8.30
cffi==1.17.1
channels==3.0.0
charset-normalizer==3.4.0
constantly==23.10.4
cryptography==43.0.3
daphne==3.0.0
Django==3.1.4
django-cors-headers==3.11.0
djangorestframework==3.12.2
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pyOpenSSL==24.2.1
pytz==2024.2
requests==2.32.3
service-identity==24.1.0
setuptools==75.2.0
sniffio==1.3.1
sqlparse==0.5.1
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
urllib3==2.2.3
zope.interface==7.1.0
